requirements.txt
# LLM Council - Multi-Agent Discussion Framework
# Compatible with Python 3.8+ including 3.13

# Paid API clients (latest versions with Python 3.13 support)
openai>=1.30.1
anthropic>=0.25.0
google-generativeai>=0.5.0
mistralai>=0.1.8

# Free/Open Source API clients (optional)
ollama>=0.1.7
groq>=0.4.2
huggingface-hub>=0.20.3

# Utilities (Python 3.13 compatible)
python-dotenv>=1.0.0
pydantic>=2.9.0
colorama>=0.4.6
rich>=13.7.1

# Optional: For enhanced features
tiktoken>=0.6.0
requests>=2.31.0

# Optional: LangChain integration (if needed)
# langchain>=0.1.20
# langchain-openai>=0.1.6
# langchain-anthropic>=0.1.11
# langchain-google-genai>=1.0.3
# langchain-mistralai>=0.1.0
# crewai>=0.28.0
# crewai-tools>=0.1.6


